requirements.txt
flask
flask_socketio
black
flask-wtf
mysql_connector